requirements.txt
requests==2.18.4
lxml==3.8.0
regex==2018.2.21
python-dateutil==2.6.1
numpy==1.14.1
gspread==0.6.2
oauth2client==4.1.2
robobrowser==0.5.3
schedule==0.4.3